+++ tests/testthat/test-field_types.R
@@ -185,4 +185,21 @@ test_that(".default_field_type must be a valid field_type", {
   )
 })
 
+test_that(".default_field_type cannot be a timepoint or strata field_type", {
+  expect_error(
+    field_types(
+      Col_tp1 = ft_numeric(),
+      .default_field_type = ft_timepoint()
+    ),
+    class = "invalid_field_types"
+  )
 
+  expect_error(
+    field_types(
+      Col_tp1 = ft_timepoint(),
+      .default_field_type = ft_strata()
+    ),
+    class = "invalid_field_types"
+  )
+
+})